Inline test module
