+++ crates/shielder-relayer/src/price_feed/mod.rs
@@ -13,8 +13,6 @@ use time::OffsetDateTime;
 use tokio::time::Duration;
 
 mod fetching;
-#[cfg(test)]
-mod tests;
 
 /// A collection of prices for various coins.
 ///
@@ -89,5 +87,39 @@ pub async fn start_price_feed(prices: Prices) -> Result<(), anyhow::Error> {
     loop {
         prices.update().await?;
         tokio::time::sleep(prices.refresh_interval).await;
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[tokio::test]
+    async fn single_update() {
+        let prices = Prices::new(Duration::from_secs(1_000_000), Default::default());
+        prices.update().await.unwrap();
+        for coin in Coin::iter() {
+            assert!(prices.price(coin).is_some());
+        }
+    }
+
+    #[tokio::test]
+    async fn with_short_validity_even_after_update_there_is_no_price_available() {
+        let prices = Prices::new(Duration::from_millis(1), Default::default());
+        prices.update().await.unwrap();
+        for coin in Coin::iter() {
+            assert!(prices.price(coin).is_none());
+        }
+    }
+
+    #[tokio::test]
+    async fn start_price_feed_works() {
+        let prices = Prices::new(Duration::from_secs(1_000_000), Duration::from_secs(1));
+        tokio::spawn(start_price_feed(prices.clone()));
+
+        tokio::time::sleep(Duration::from_secs(10)).await;
+        for coin in Coin::iter() {
+            assert!(prices.price(coin).is_some());
+        }
     }
 }
